omnifoundry/core/config.py
"""
配置管理模組

負責載入和管理系統配置。
"""

import yaml
from pathlib import Path


class Config:
    """配置管理器"""
    
    def __init__(self, config_path=None):
        self.config_path = config_path
        self.config = {}
    
    def load(self):
        """載入配置檔"""
        pass
    
    def save(self, path=None):
        """儲存配置檔"""
        pass
    
    def get(self, key, default=None):
        """獲取配置值"""
        pass
    
    def set(self, key, value):
        """設定配置值"""
        pass

