tests/test_model_manager.py
"""
模型管理模組測試
"""

import pytest
from omnifoundry.core.model_manager import ModelManager


class TestModelManager:
    """模型管理器測試"""
    
    def test_init(self):
        """測試初始化"""
        manager = ModelManager()
        assert manager is not None
        assert manager.cache_dir == "./models"
    
    def test_search_models(self):
        """測試模型搜尋"""
        # TODO: 實作測試
        pass
    
    def test_download_model(self):
        """測試模型下載"""
        # TODO: 實作測試
        pass
    
    def test_list_local_models(self):
        """測試列出本地模型"""
        # TODO: 實作測試
        pass

